# backend/src/modules/ai-integration/openai/gpts-openapi-schema.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Agentrix Marketplace API",
    "version": "1.0.2",
    "description": "Agentrix Marketplace API for GPTs. Supports product search, product details, order creation, and payment initiation."
  },
  "servers": [
    {
      "url": "https://api.agentrix.top/api",
      "description": "Agentrix API Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API Key for identifying the GPT user session."
      }
    },
    "schemas": {
...
      "ProductResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "message": { "type": "string" },
          "data": { "$ref": "#/components/schemas/Product" }
        },
        "required": ["success", "data"]
      },
      "Product": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "description": "Product ID" },
          "title": { "type": "string", "description": "Product title" },
          "description": { "type": "string" },
          "price": { "type": "number" },
          "currency": { "type": "string", "default": "CNY" },
          "priceDisplay": { "type": "string" },
          "image": { "type": "string", "format": "uri" },
          "images": {
            "type": "array",
            "items": { "type": "string", "format": "uri" }
          },
          "category": { "type": "string" },
          "productType": {
            "type": "string",
            "enum": ["physical", "service", "crypto", "nft"]
          },
          "stock": { "type": "integer" },
          "inStock": { "type": "boolean" }
        },
        "required": ["id", "title", "price"]
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "error": { "type": "string" },
          "message": { "type": "string" }
        },
        "required": ["success"]
      }
    }
  },
  "security": [{ "ApiKeyAuth": [] }],
  "paths": {
...
    "/marketplace/products/{id}": {
      "get": {
        "summary": "Get product details",
        "operationId": "getProduct",
        "tags": ["Marketplace"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Product details",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ProductResponse" }
              }
            }
          },
          "404": {
            "description": "Product not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
...
  }
}